refactor: 将webhook配置验证移到前端
