+++ src/one_dragon_qt/view/setting/setting_push_interface.py
@@ -1,3 +1,5 @@
+import json
+
 from PySide6.QtWidgets import QWidget
 from qfluentwidgets import FluentIcon, PushButton, InfoBar, InfoBarPosition
 
@@ -157,6 +159,13 @@ class SettingPushInterface(VerticalScrollInterface):
         """发送测试消息到当前选择的通知方式"""
         selected_method = self.notification_method_opt.getValue()
         test_method = str(selected_method)
+        if test_method == "WEBHOOK":
+            # 如果是Webhook方式，先验证配置
+            try:
+                self._validate_webhook_config()
+            except ValueError as e:
+                self._show_error_message(str(e))
+                return
 
         try:
             pusher = Push(self.ctx)
@@ -254,6 +263,42 @@ class SettingPushInterface(VerticalScrollInterface):
         # 复制到剪贴板
         PcClipboard.copy_string(curl_command)
         self._show_success_message("cURL 命令已复制到剪贴板！")
+
+    def _validate_webhook_config(self) -> None:
+        """
+        验证Webhook配置
+        验证失败时抛出异常
+        """
+        url = getattr(self.ctx.push_config, "webhook_url", None)
+        if not url:
+            raise ValueError("Webhook URL 未配置，无法推送")
+
+        body = getattr(self.ctx.push_config, "webhook_body", None)
+        headers = getattr(self.ctx.push_config, "webhook_headers", "{}")
+        content_type = getattr(self.ctx.push_config, "webhook_content_type", "application/json")
+
+        # 检查是否包含 $content
+        if not any('$content' in str(field) for field in [url, body, headers]):
+            raise ValueError("URL、请求头或者请求体中必须包含 $content 变量")
+
+        # 如果是JSON格式，验证JSON的合法性
+        if content_type == "application/json":
+            # 检查body模板是否为合法JSON
+            if not self._validate_json_format(body):
+                raise ValueError("请求体不是合法的JSON格式")
+
+        # 检查请求头是否为合法JSON
+        if headers and headers != "{}":
+            if not self._validate_json_format(headers):
+                raise ValueError("请求头不是合法的JSON格式")
+
+    def _validate_json_format(self, json_str: str) -> bool:
+        """验证JSON格式的合法性"""
+        try:
+            json.loads(json_str)
+            return True
+        except (json.JSONDecodeError, TypeError):
+            return False
 
     def _show_success_message(self, message: str):
         """显示成功消息提示"""
